Update modal-header.component.ts
According to WCAG and Accessibility Checker plugin, buttons require "visible" text. Because of this, `aria-label` is insufficient... Adding the text as, but visually hiding it meets all a11y requirements, and the `aria-label` is not longer necessary.
Voiceover and JAWS will read the button label when they announce the button element, using this approach.