fix(ibm-dropdown-list): ensure ListItem.disabled is treated as a boolean by styling logic
The disabled styling was treating ListItem.disabled input as a disabled attribute (i.e. present or
not) even though the doc for ListItem says that it is a boolean. Now, we treat ListItem.disabled
input as a boolean.

Add tests for disabling a list-item

Add disabled item examples to dropdown and combobox storybooks

Fix linting error in grid

Signed-off-by: anemonetea <laz1anastasiya@gmail.com>